Bug fix - wrong behaviour for should_create_index
The `should_create_index` was implemented like it means `do_create_index`, instead of behaving like `allow_existing_index`, which was the original purpose.

When the user is setting `should_create_index=False`, he doesn't necessarily mean "**don't** create index", but rather "it's ok if the index exists, I allow appending to it".

The confusion is probably in the naming - we should have called this argument `allow_existing_index` to begin with. Let's consider renaming it in the next major release